create new run types package and move reflection files there
