+++ packages/router/tsconfig.build-esm.json
@@ -6,5 +6,6 @@
   },
   "reflection": true,
   "include": ["."],
-  "exclude": ["node_modules", "**/*.spec.ts", "examples", ".dist"]
+  "exclude": ["node_modules", "**/*.spec.ts", "examples", ".dist"],
+  "references": [{"path": "../runtype"}]
 }
