+++ packages/runtype/index.ts
@@ -1,0 +1,10 @@
+/* ########
+ * 2022 mion
+ * Author: Ma-jerez
+ * License: MIT
+ * The software is provided "as is", without warranty of any kind.
+ * ######## */
+
+export * from './src/constants';
+export * from './src/reflection';
+export * from './src/types';
