+++ packages/runtype/src/constants.ts
@@ -1,0 +1,32 @@
+/* ########
+ * 2023 mion
+ * Author: Ma-jerez
+ * License: MIT
+ * The software is provided "as is", without warranty of any kind.
+ * ######## */
+
+import {ReflectionOptions} from './types';
+
+/** Reflection and Deepkit Serialization-Validation options */
+export const DEFAULT_REFLECTION_OPTIONS: Readonly<ReflectionOptions> = {
+    /**
+     * Deepkit Serialization Options
+     * loosely defaults to false, Soft conversion disabled.
+     * !! We Don't recommend to enable soft conversion as validation might fail
+     * */
+    serializationOptions: {
+        loosely: false,
+    },
+
+    /**
+     * Deepkit custom serializer
+     * @link https://docs.deepkit.io/english/serialization.html#serialisation-custom-serialiser
+     * */
+    customSerializer: undefined,
+
+    /**
+     * Deepkit Serialization Options
+     * @link https://docs.deepkit.io/english/serialization.html#_naming_strategy
+     * */
+    serializerNamingStrategy: undefined,
+};
