+++ packages/runtype/tsconfig.build.json
@@ -5,6 +5,6 @@
   },
   "reflection": true,
   "include": ["."],
-  "exclude": ["examples", ".dist"],
-  "references": [{"path": "../runtype"}]
+  "exclude": ["node_modules", "**/*.spec.ts", "examples", ".dist"],
+  "references": []
 }
