+++ packages/runtype/tsconfig.json
@@ -4,7 +4,7 @@
     "outDir": ".dist/cjs"
   },
   "reflection": true,
-  "include": ["."],
-  "exclude": ["examples", ".dist"],
-  "references": [{"path": "../runtype"}]
+  "include": [""],
+  "exclude": [".dist"],
+  "references": []
 }
